Lua build fixups for Posix.
Since we build directly in lib/lua/src, the Makefile requires $(PLAT)
be set to posix.

Also, set $(CC) to 'cc', remove -Wall, and change -O2 to -O.  This
should let us build on anything even remotely POSIX compliant.

At some point we might need to figure out a way to have the lua
library build pay attention to configure's '--debug' flag. But
that's only required if we have a need to actually step through the
Lua library source code in a debugger, and that's not a likely
requirement.